fixed a couple of warnings

+++ Libraries/snappy/snappy.cc
@@ -1011,7 +1011,7 @@ class SnappyIOVecWriter {
       assert(curr_iov_written_ <= output_iov_[curr_iov_index_].iov_len);
       if (curr_iov_written_ >= output_iov_[curr_iov_index_].iov_len) {
         // This iovec is full. Go to the next one.
-        if (curr_iov_index_ + 1 >= output_iov_count_) {
+        if (size_t(curr_iov_index_ + 1) >= output_iov_count_) {
           return false;
         }
         curr_iov_written_ = 0;
@@ -1093,7 +1093,7 @@ class SnappyIOVecWriter {
                                   len);
         if (to_copy == 0) {
           // This iovec is full. Go to the next one.
-          if (curr_iov_index_ + 1 >= output_iov_count_) {
+          if (size_t(curr_iov_index_ + 1) >= output_iov_count_) {
             return false;
           }
           ++curr_iov_index_;
